Add extract CI from all headers and partial file for readability
